fix: constant로 로직 변경
