+++ src/components/ProjectAbout/PR/PRCard.tsx
@@ -1,6 +1,7 @@
-import { IPRRequestType, PRType } from '@/api/pr-records/type';
+import { IPRRequest, PRType } from '@/api/pr-records/type';
 import * as _ from './PRCard.style';
 import { VStack } from '@/components/Stack';
+import { PRConstant } from '@/constant/PRType';
 
 const PRCard = ({
   title,
@@ -8,29 +9,16 @@ const PRCard = ({
   content,
   solution,
   attachmentUrls,
-}: IPRRequestType) => {
-  let labelName = '';
-
-  switch (type) {
-    case PRType.NEW_FEATURE:
-      labelName = '기능설명';
-      break;
-    case PRType.BUG_FIX:
-      labelName = '문제사항';
-      break;
-    case PRType.REFACTORING:
-      labelName = '개선부분';
-      break;
-  }
+}: IPRRequest) => {
 
   return (
     <_.PRCardWrapper>
       <_.PRTitle>{title}</_.PRTitle>
       <VStack gap={10}>
-        <_.SubTitle>{labelName}</_.SubTitle>
+        <_.SubTitle>{PRConstant[type].label[0]}</_.SubTitle>
         <_.Content>{content}</_.Content>
       </VStack>
-      {type == PRType.BUG_FIX && solution && (
+      {type == 'BUG_FIX' && solution && (
         <VStack gap={10}>
           <_.SubTitle>해결방법</_.SubTitle>
           <_.Content>{solution}</_.Content>
