+++ src/components/ProjectAbout/PR/index.tsx
@@ -1,9 +1,9 @@
-import {  IPRRequestListType } from '@/api/pr-records/type';
+import {  IPRRequestList } from '@/api/pr-records/type';
 import * as Common from '../Common/style';
 import * as _ from './style';
 import PRCard from './PRCard';
 
-const PRList = ({ prRecords }: IPRRequestListType) => {
+const PRList = ({ prRecords }: IPRRequestList) => {
   return (
     <Common.ContentWrapper>
       <Common.Label>PR 기록</Common.Label>
