+++ src/components/ProjectWriting/PR/Modal/index.tsx
@@ -1,30 +1,82 @@
-import { IPRRecordsType, PRType } from '@/api/pr-records/type';
+import { IPRRecords } from '@/api/pr-records/type';
 import * as _ from './Modal.style';
-import { FeatModal } from './FeatModal';
-import { BugFixModal } from './BugFixModal';
-import { RefactModal } from './RefactModal';
+import { HStack } from '@/components/Stack';
+import { TextAreaInput } from '@/components/Input';
+import { DeleteIcon } from '@/asset/icon';
+import { GetPRContent, PatchPRContent } from '@/api/pr-records';
+import { useEffect, useState } from 'react';
+import { Button } from '@/components/Button';
+import { PRConstant } from '@/constant/PRType';
 
-export interface ModalPropsType {
-  pr: IPRRecordsType;
+export interface IModalProps {
+  pr: IPRRecords;
   onClose: () => void;
-};
+};  
 
-const Modal = ({ pr, onClose }: ModalPropsType) => {
-  const PRMoalSwitch = () => {
-    if (pr) {
-      switch (pr.type) {
-        case PRType.NEW_FEATURE:
-          return <FeatModal pr={pr} onClose={onClose} />;
-        case PRType.BUG_FIX:
-          return <BugFixModal pr={pr} onClose={onClose} />;
-        case PRType.REFACTORING:
-          return <RefactModal pr={pr} onClose={onClose} />;
+const Modal = ({ pr, onClose }: IModalProps) => {
+
+    const [content, setContent] = useState<string>('');
+    const [solution, setSolution] = useState<string | undefined>('');
+
+    const { data: PRContent } = GetPRContent(
+      pr.id
+    );
+    const { mutate: PRContentMutation } = PatchPRContent(pr.id);
+    const [images, setImages] = useState<string[]>([]);
+
+    useEffect(() => {
+      if (PRContent) {
+        setContent(PRContent.content);
+        setSolution(PRContent?.solution);
+        setImages(PRContent.attachmentUrls);
       }
-    }
-  };
+    }, [PRContent]);
+
+    const onPatch = () => {
+      PRContentMutation({
+        title: pr.title,
+        importance: pr.importance,
+        type: pr.type,
+        content: content,
+        solution: solution,
+        attachmentUrls: images,
+      });
+    };
+
   return (
     <_.ModalBackground>
-      <PRMoalSwitch />
+      <_.PRModalWrapper>
+        <_.ModalHeaderWrapper>
+          <HStack gap={30}>
+            <_.ModalLabel>{pr.title}</_.ModalLabel>
+            <_.TypeBox>{PRConstant[pr.type].typeName}</_.TypeBox>
+          </HStack>
+          <DeleteIcon onClick={onClose} />
+        </_.ModalHeaderWrapper>
+        <_.ModalMainWrapper>
+          <TextAreaInput
+            isAddImage={true}
+            isMapImage={pr.type !== "BUG_FIX"}
+            images={images}
+            setImages={setImages}
+            value={content}
+            onChange={(e) => setContent(e.target.value)}
+            label={PRConstant[pr.type].label[0]}
+          />
+          {pr.type === 'BUG_FIX' && (
+            <TextAreaInput
+              isAddImage={true}
+              isMapImage={true}
+              images={images}
+              setImages={setImages}
+              value={solution}
+              onChange={(e) => setSolution(e.target.value)}
+              label={PRConstant.BUG_FIX.label[1]}
+            />
+          )}
+          <Button onClick={onPatch}>저장</Button>
+        </_.ModalMainWrapper>
+      </_.PRModalWrapper>
     </_.ModalBackground>
   );
 };
