+++ src/constant/PRType.ts
@@ -1,0 +1,5 @@
+export const PRConstant = {
+  NEW_FEATURE: { typeName: '기능 추가', label: ['기능 설명'] },
+  BUG_FIX: { typeName: '버그 수정', label: ['버그 설명', '버그 해결방법'] },
+  REFACTORING: { typeName: '기능 추가', label: ['개선 부분'] },
+};
